Billing - Checkout
[Billing]
- Payments for Client can now be fetched

[Tables]
- Re-assigned data types of costs and pricing from integer -> float.

+++ database/migrations/2007_04_19_001232_create_supplies_mstr_table.php
@@ -14,8 +14,8 @@ class CreateSuppliesMstrTable extends Migration
             $table->integer('category_id');
             $table->char('uom', 10);
             $table->integer('uom_qty');
-            $table->integer('selling_price');
-            $table->integer('cost');
+            $table->float('selling_price', 6, 2);
+            $table->float('cost', 8, 2);
         });
     }
